migrate JS tracking from WP AJAX to WP REST API
Use a custom WP REST route for tracking requests. The tracking
logic is moved up into Statify class. The API logic is separated
into a new class, so additional endpoints can be added there.

Override the default nonce behavior for the tracking endpoint to
preserve the logged-in flag with nonce verification disabled.